Add ability to scan external IP without proxy file
Adds a new ProxyFile option to specify filename containing proxies.
If it doesn't exist, instead of aborting we add the option to skip
proxies entirely and check the actual IP for bans instead.

A nice idea in future would be a command line flag to force checking
the external IP and completely skip checks for a proxy file.